# Mixer standalone perf benchmark tests