--- a/mixer/pkg/perf/clientserver.go
+++ b/mixer/pkg/perf/clientserver.go
@@ -134,11 +134,17 @@ func (s *ClientServer) close() (err error) {
 // ClientServerInitParams is a collection of parameters that are passed as part of the InitializeClient call.
 type ClientServerInitParams struct {
 
-	// Setup is the YAML-serialized setup object.
-	Setup []byte
+	// Setup is the YAML-serialized load object.
+	Load []byte
 
 	// Address of the Mixer Server.
 	Address string
+
+	// IdentityAttribute for Mixer server
+	IdentityAttribute string
+
+	// IdentityAttributeDomain for the request
+	IdentityAttributeDomain string
 }
 
 // InitializeClient is a remote RPC call that is invoked by the controller to initiate setup of the client environment.
@@ -148,11 +154,11 @@ func (s *ClientServer) InitializeClient(params ClientServerInitParams, _ *struct
 	log.Infof("ClientServer initializing with server address: %s", params.Address)
 	_ = log.Sync()
 
-	var setup Setup
-	if err := unmarshallSetup(params.Setup, &setup); err != nil {
+	var load Load
+	if err := unmarshallLoad(params.Load, &load); err != nil {
 		return err
 	}
-	return s.client.initialize(params.Address, &setup)
+	return s.client.initialize(params.Address, &load, params.IdentityAttribute, params.IdentityAttributeDomain)
 }
 
 // Shutdown is a remote RPC call that is invoked by the controller after the benchmark execution has completed.