--- a/mixer/pkg/perf/run.go
+++ b/mixer/pkg/perf/run.go
@@ -137,7 +137,6 @@ func run(b benchmark, setup *Setup, settings *Settings, coprocess bool) {
 	b.run(name, func(bb benchmark) {
 		_ = controller.runClients(bb.n())
 	})
-	b.logf("completed running test: %s", b.name())
 
 	// Even though we have a deferred close for controller, do it explicitly before leaving control to perform
 	// graceful close of clients during teardown.
@@ -168,7 +167,12 @@ func runDispatcherOnly(b benchmark, setup *Setup, settings *Settings) {
 		globalDict[list[i]] = int32(i)
 	}
 
-	requests := setup.Load.createRequestProtos(setup.Config)
+	// there has to be just 1 load for InProcessBypassGrpc case
+	if len(setup.Loads) != 1 {
+		b.fatalf("for `InProcessBypassGrpc`, load must contain exactly 1 entry")
+		return
+	}
+	requests := setup.Loads[0].createRequestProtos(setup.Config.IdentityAttribute, setup.Config.IdentityAttributeDomain)
 	bags := make([]attribute.Bag, len(requests)) // precreate bags to avoid polluting allocation data.
 	for i, r := range requests {
 		switch req := r.(type) {