Added the following perf tests for Mixer grpc server. The tests execute the call from mixer client and invoke the Mixer's grpc endpoint. Here are the [performance numbers](https://github.com/guptasu/istio/blob/tryPerf/mixer/test/perf/bench.baseline)

* Also 
  * enhanced the spy adapters to all introducing latency so that we can simulate over the network adapter
  * enhanced the perf library to execute clients in parallel with varying load. Did some minor cleanup.
  * added support for quota template in the spy adapter.

Interesting files to review are:
1. [check_test.go](https://github.com/istio/istio/compare/master...guptasu:tryPerf?expand=1#diff-76660409077208ecd184976030ff95fa)
2. [quota_test.go](https://github.com/istio/istio/compare/master...guptasu:tryPerf?expand=1#diff-9f9052503df0d0076a6f3351f07351f8)
3. [report_test.go](https://github.com/istio/istio/compare/master...guptasu:tryPerf?expand=1#diff-094506c3359f9a2d0fe01c910ff939ba)

- Tests
# 05/15/2018
# content of this file is generated by running `~/go/src/istio.io/istio/mixer/test/perf$ go test -bench=. -benchmem`
# report calls
Benchmark_Report_1Client_1Call/InProc-12                                   	    5000	    278474 ns/op	   31664 B/op	     483 allocs/op
Benchmark_Report_1Client_5SameCalls/InProc-12                              	    1000	   1442520 ns/op	  160997 B/op	    2452 allocs/op
Benchmark_Report_1Client_5DifferentCalls/InProc-12                         	    1000	   1394915 ns/op	  161518 B/op	    2464 allocs/op
Benchmark_Report_4Clients_5SameCallsEach/InProc-12                         	    1000	   1470807 ns/op	  164845 B/op	    2502 allocs/op
Benchmark_Report_4Clients_5DifferentCallsEach/InProc-12                    	    1000	   1437547 ns/op	  164593 B/op	    2508 allocs/op
Benchmark_Report_4Clients_5SameCallsEach_1MilliSecSlowApa/InProc-12        	     200	   6972848 ns/op	  251941 B/op	    3734 allocs/op
# quota calls
Benchmark_Quota_1Client_1Call/InProc-12                                    	    5000	    267662 ns/op	   33692 B/op	     361 allocs/op
Benchmark_Quota_1Client_5SameCalls/InProc-12                               	    1000	   1337831 ns/op	  168765 B/op	    1813 allocs/op
Benchmark_Quota_1Client_5DifferentCalls/InProc-12                          	    1000	   1334904 ns/op	  169832 B/op	    1828 allocs/op
Benchmark_Quota_4Clients_5SameCallsEach/InProc-12                          	    1000	   1346433 ns/op	  170895 B/op	    1843 allocs/op
Benchmark_Quota_4Clients_5DifferentCallsEach/InProc-12                     	    1000	   1348580 ns/op	  171978 B/op	    1858 allocs/op
Benchmark_Quota_4Clients_5SameCallsEach_1MilliSecSlowApa/InProc-12         	     200	   6898903 ns/op	  217486 B/op	    2488 allocs/op
# check calls
Benchmark_Check_1Client_1Call/InProc-12                         	    5000	    262751 ns/op	   26265 B/op	     285 allocs/op
Benchmark_Check_1Client_5SameCalls/InProc-12                    	    1000	   1234255 ns/op	  131710 B/op	    1434 allocs/op
Benchmark_Check_1Client_5DifferentCalls/InProc-12               	    1000	   1237198 ns/op	  132444 B/op	    1445 allocs/op
Benchmark_Check_4Clients_5SameCallsEach/InProc-12               	    1000	   1246696 ns/op	  133226 B/op	    1456 allocs/op
Benchmark_Check_4Clients_5DifferentCallsEach/InProc-12          	    1000	   1248183 ns/op	  134033 B/op	    1467 allocs/op
Benchmark_Check_4Clients_5SameCallsEach_1MilliSecSlowApa/InProc-12         	     200	   6785091 ns/op	  152366 B/op	    1721 allocs/op