@istio-testing 2018-05-15
[APPROVALNOTIFIER] This PR is **NOT APPROVED**

This pull-request has been approved by: 
To fully approve this pull request, please assign additional approvers.
We suggest the following additional approver: **douglas-reid**

Assign the PR to them by writing `/assign @douglas-reid` in a comment when ready.

The full list of commands accepted by this bot can be found [here](https://go.k8s.io/bot-commands).

The pull request process is described [here](https://git.k8s.io/community/contributors/guide/owners.md#the-code-review-process)

<details open>
Needs approval from an approver in each of these files:

- **[bin/OWNERS](https://github.com/istio/istio/blob/master/bin/OWNERS)**
- **[mixer/OWNERS](https://github.com/istio/istio/blob/master/mixer/OWNERS)**

Approvers can indicate their approval by writing `/approve` in a comment
Approvers can cancel approval by writing `/approve cancel` in a comment
</details>
<!-- META={"approvers":["douglas-reid"]} -->