@istio-testing 2018-05-16
@guptasu: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
prow/istio-unit-tests.sh | 9419a65a90241ae42807ff7392677c3537706578 | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_istio/5619/istio-unit-tests/7407/) | `/test istio-unit-tests`
prow/istio-pilot-e2e-envoyv2-v1alpha3.sh | 9419a65a90241ae42807ff7392677c3537706578 | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_istio/5619/istio-pilot-e2e-envoyv2-v1alpha3/1562/) | `/test istio-pilot-e2e-envoyv2-v1alpha3`
prow/istio-pilot-e2e.sh | 9419a65a90241ae42807ff7392677c3537706578 | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_istio/5619/istio-pilot-e2e/7189/) | `/test istio-pilot-e2e`
prow/e2e-bookInfoTests-v1alpha3.sh | 9419a65a90241ae42807ff7392677c3537706578 | [link](https://k8s-gubernator.appspot.com/build/istio-prow/pull/istio_istio/5619/e2e-bookInfoTests-envoyv2-v1alpha3/1022/) | `/test e2e-bookInfo-envoyv2-v1alpha3`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/devel/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->